+++ packages/linode-js-sdk/src/nodebalancers/index.ts
@@ -1,0 +1,9 @@
+export * from './types';
+
+export * from './nodebalancers.schema';
+
+export * from './nodebalancers';
+
+export * from './nodebalancer-configs';
+
+export * from './nodebalancer-config-nodes';
