+++ packages/linode-js-sdk/tslint.json
@@ -7,6 +7,7 @@
     "interface-name": false,
     "object-literal-sort-keys": false,
     "ban-types": false,
-    "variable-name": false
+    "variable-name": false,
+    "array-type": false
   }
 }
