+++ packages/manager/src/services/nodebalancers/types.ts
@@ -1,0 +1,10 @@
+import {
+  NodeBalancer,
+  NodeBalancerConfigPort
+} from 'linode-js-sdk/lib/nodebalancers';
+
+export interface ExtendedNodeBalancer extends NodeBalancer {
+  up: number;
+  down: number;
+  configPorts: NodeBalancerConfigPort[];
+}
